modules/libpr0n/test/mochitest/test_bug468160.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=468160
-->
<head>
  <title>Test for Bug 468160</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=468160">Mozilla Bug 468160</a>
<p id="display"></p>
<div id="content" style="display: none">
<!-- bug468160.sjs does a simple redirect to a PNG image. As long as this
     doesn't leak, this test passes. -->

<object data="bug468160.sjs"></object>
  
</div>
<pre id="test">
<script type="application/javascript">

ok(true, "test passed");

</script>
</pre>
</body>
</html>
